// source/properties.d.ts
/**
 * Copyright (C) 2018 Silas B. Domingos
 * This source code is licensed under the MIT License as described in the file LICENSE.
 */

/**
 * Import properties interface.
 */
export interface Properties {
  /**
   * Import classes.
   */
  class?: string;
  /**
   * Import slot.
   */
  slot?: string;
  /**
   * Import path.
   */
  path?: string;
  /**
   * Import children.
   */
  children?: {};
}